refactor(web): co-locate route components
