+++ apps/web/src/app/(dashboard)/cars/_components/makes/make-detail.tsx
@@ -1,6 +1,6 @@
+import { MakeSelector } from "@web/app/(dashboard)/cars/_components/make-selector";
+import { MakeTrendChart } from "@web/app/(dashboard)/cars/_components/makes";
 import type { Logo } from "@web/app/(dashboard)/cars/makes/[make]/page";
-import { MakeSelector } from "@web/components/cars/make-selector";
-import { MakeTrendChart } from "@web/components/cars/makes";
 import { LastUpdated } from "@web/components/shared/last-updated";
 import NoData from "@web/components/shared/no-data";
 import { columns } from "@web/components/tables/columns/cars-make-columns";
