+++ apps/web/src/app/(dashboard)/coe/_components/coe-categories.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CategoryInfo } from "@web/components/cars/category-info";
+import { CategoryInfo } from "@web/app/(dashboard)/cars/_components/category-info";
 import {
   Card,
   CardContent,
