+++ apps/web/src/app/(dashboard)/coe/_components/latest-coe-prices.tsx
