+++ apps/web/src/app/(dashboard)/coe/_components/pqp/animated-section.tsx
