+++ apps/web/src/app/(dashboard)/coe/_components/pqp/comparison-summary-card.tsx
