+++ apps/web/src/app/(dashboard)/coe/_components/pqp/index.ts
