+++ apps/web/src/app/(dashboard)/coe/categories/[category]/page.tsx
@@ -1,10 +1,10 @@
 import { redis } from "@sgcarstrends/utils";
+import { COEPremiumChart } from "@web/app/(dashboard)/coe/_components/premium-chart";
 import {
   getDefaultEndDate,
   getDefaultStartDate,
   loadSearchParams,
 } from "@web/app/(dashboard)/coe/search-params";
-import { COEPremiumChart } from "@web/components/coe/premium-chart";
 import { PageHeader } from "@web/components/page-header";
 import { StructuredData } from "@web/components/structured-data";
 import Typography from "@web/components/typography";
