+++ apps/web/src/app/blog/_components/progress-bar.tsx
