+++ apps/web/src/components/category-info.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
+import { CategoryInfo } from "@web/app/(dashboard)/cars/_components/category-info";
 import { Car } from "lucide-react";
-import { CategoryInfo } from "./cars/category-info";
 
 const mockCategories = {
   "Category A": true,
